- [*] split examples_test.go to example_test & example0_test.go
