+++ example_test.go
@@ -1,0 +1,25 @@
+package easygen_test
+
+import (
+	"os"
+
+	"github.com/go-easygen/easygen"
+	"github.com/go-easygen/easygen/egVar"
+)
+
+// for standalone test, change package to `main` and the next func def to,
+// func main() {
+func ExampleProcess() {
+	tmpl0 := easygen.NewTemplate().Customize()
+	tmpl := tmpl0.Funcs(easygen.FuncDefs()).Funcs(egVar.FuncDefs())
+	easygen.Process(tmpl, os.Stdout, "test/var0")
+
+	// Output:
+	// Input: "some-init-method"
+	// Output 1: "SomeInitMethod"
+	// Output 2: "SOME_INIT_METHOD"
+}
+
+// To show the full code in GoDoc
+type dummy struct {
+}
